Fix #232 Allow media playback to work when device locked or muted
Added support for AVAudioSession to allow media files to continue playing when device is locked or muted. This is a change in behavior from 1.0 where playback would pause when the device was locked or could not be heard when the device was muted.
Recording behavior is not changed.  Recording will occur if the mute button is set and will continue recording if the device is locked during recording.
Sending the app to the background will pause playing or recording and resume when the app is returned to the foreground (depending upon the length of the audio).  If you want to stop playing or recording when the app enters the background listen for the pause event and call your media object's stop() or stopRecord() functions.