Add OpenAllWhitelistURLsInWebView policy setting to determine whether ALL whitelisted URLs are opened within the embedded web view (as is sometimes desired).
For some applications it is desirable that ALL whitelisted URLs be opened within the embedded web view, regardless of the navigation mode that caused the URL to be loaded (for instance target=_blank).  This setting adds the ability for devs to tailor the URL loading policy to match their apps needs.