Fixes CB-819 fail callback not invoked
Problem was that on iPad when popover controller is used,
the fail callback was not getting called when the popover was
dismissed w/o picking an image.  Added a delay to the callback to JavaScript
since the app will hang if the callback contains a JavaScript alert.
Also removed clause to close popover from imagePickerDidCancel
because that method is not called when the popover is used.
Also deprecated closePicker since it is not
used within the Cordova code.  Deprecated in the off chance a plugin
is using - will remove in 2.0.