Change wire format of exec handler.
The JS now sends commands as a JSON array instead of pre-processed JSON
object. The code maintains compatibility with old-style plugins by
performing the "options" dict extraction on the native side when the
plugin responds to method:withDict: selector. The new way of
implementing native callbacks is to us a method called method:, which is
passed the CDVInvokedUrlCommand.

Update CDVFileTransfer to use the new API.