Fix CB-2271 - Multiple Cordova Views.
The problem was that multiple views loading at the same time would end
up with the same User-Agent. This change adds a locking mechanism around
User-Agent setting, so that multiple views will load in serial instead
at the same time.

Adds a unit test that failed before the fix, but now passes consistently

https://issues.apache.org/jira/browse/CB-2271