[CB-1108] Convert <plugin> -> <feature> with <param>
This unifies the config.xml used by iOS and Android. Now <feature> tags can be used
to add plugins. The changes are non-intrusive so the older versions of the config.xml
can still be used after this change. The value of a param named "ios-package"
determines the class for the plugin. Also changes the config.xml with the unified one
on the project template