Allow the start page to be loaded from a file URL
This change allows the start page to be loaded from any location
inside the app (e.g the Documents or Caches directory). Currently
it can only load the start page from the app's bundle directory.
Because [NSURL urlWithString:...] does not work with file URLS,
there needs to be a mechanism to specify that wwwFolderName is a
file URL.