Remove duplicate reference to the same libCordova.a.
libCordova.a has already been referred to as a framework, no need to
force_load it again.

Normally it is OK to refer to libCordova.a twice, as they are the same
file anyway. But when symbolic link is used in workspace and Jenkins is
use to do the build, there will be a problem because Jenkins will
dereference symbolic links (https://issues.jenkins-ci.org/browse/JENKINS-5597)
and now we could have 2 copies of libCordova.a, and have countless
duplicate symbols errors.