CB-12130 - Launch storyboard images are not updated or cleaned
The platform project directory needs to be relative to the project

Launch screen images were not updated or cleaned in
`updateLaunchStoryboardImages()` or `cleanLaunchStoryboardImages()`
because the check for `Images.xcassets` in
`getLaunchStoryboardImagesDir()` failed.

The reason was that the `platformProjDir` and `projectRoot` were
joined, while both were relative to the root directory. So the
resulting directory was non-sensical. This changes makes
`platformProjDir` relative to `cordovaProject.root`, which matches the
same check in `updateSplashScreens()` and `updateIcons()`.

 This closes #276

(cherry picked from commit 99219c1c4cd8d1000fb89a15449f2d25a57bef41)