CB-12084: Update project build settings & plist
If no legacy images are provided, we should remove the launch image asset
compiler setting. If legacy images are present, we should make sure it is
present. We also need to make sure to properly delete launch storyboard rather
than setting to undefined.

Other fix:
Roots need to be absolute for testing to work
Contents.json should be properly capitalized

 This closes #279