Hard link instead of soft linking
Softlinking resource files (.wav files for example) leads to some unexpected behavior. Said resource files are stored as symbolic links inside the .ipa and the app cannot be installed on iOS 10.
Instead of creating symbolic link create regular links - it is better than copying the files performance-wise and the solution is more robust.