CB-13927 - Modified xcodeProjDir to filter out files/folders that contain "._" (#355)
* Modified xcodeProjDir to filter out files/folders that contain "._"

I ran into an issue trying to add an ios plugin using a virtual machine.  The code would return an array for the xcodeProjDir and then of course use the wrong one.  It would use the one starting with "._ProjectName" which I believe is an MacOS hidden folder.  I attempted to install the plugin on OsX but because it's a virtual machine it still had those hidden folders.   I've made it so that it will filter those out while looking for the proper xCode project folder.