GH-540: Fix copy-www-build-step without shelljs
There didn't seem to be a compelling reason to keep this as JavaScript
since it was primarily just shelling out to run various commands. By
keeping it in bash, we can significantly simplify the invocation in the
xcodeproj file and we can run it without a dependency on shelljs.

I also moved it into a Scripts folder inside the project so that our
xcodeproj file isn't reliant on the cordova/lib folder existing, and
this particular script isn't actually a "lib" in the standard sense.