Don't silently ignore creation of new windows
Currently, attempts to create new windows (via links with
`target="_blank"` or JS `window.open`) for pages within the application
have no effect and are silently ignored.
(Links to pages outside the application trigger the allow-intent
functionality and open in Safari.)

With this PR, the default behaviour will be to open those links in the
**same** Web View (without creating a new window).

This also introduces a preference (`AllowNewWindows`) that allows the
creation of a new UIViewController to present the new page in a new Web
View frame on top of the existing Web View.
**Note:** This new window provides no controls, so it can only be closed
with `window.close()` in JS.