Fix SplashScreen images not rendering
Trying to take a snapshot of the view doesn't seem to result in the
image actually being present in the snapshot, but we can just render the
view itself from the storyboard file.

However, we do need to make sure that we don't allow the view to grow in
size, otherwise it will expand to the size of the image instead of
remaining the size of the screen bounds.