fix(splashscreen): Don't show splashscreen when used as a lib
When using CDVViewController via CordovaLib as a library, the app might
already be loaded and not need to show the launch storyboard again.

Rather than unconditionally showing the splashscreen until the webview
content has loaded, make it initially invisible. Then in the project
template, make it visible so that it continues working as expected for
apps using the Cordova CLI.

Closes #929.