chore!: Modernize the template app (WIP)
* Restructures the Xcode project to more closely match what gets generated
  when you create a new Objective C iOS app with Xcode.
* Uses a storyboard for the main ViewController rather than a xib file.
* Renames Images.xcassets to Assets.xcassets.
* Uses CordovaLib from Swift Package Manager rather than a subproject.
* Enables Mac Catalyst support.
* Adopts SceneDelegate API on iOS 13+.

That last point might be the source of most problems, since we probably
need to plumb a bunch of stuff that used to fire on AppDelegate and now
fires on the SceneDelegate. Might need to introduce a CDVSceneDelegate
in CordovaLib to account for this.

This needs a LOT more testing before it's safe to merge 😬