fix(create): Ensure unix paths in Xcode project file
Cordova-iOS requires macOS for building and deploying, but it seems to
be a fairly common problem that people generate the iOS project files on
Windows (which isn't technically supported, but _appears_ to work) and
then Xcode can't open it on macOS because of a bad path separator.

The path separator problem is fixable, so let's at least generate a
valid Xcode project.

This does *not* claim to add support for using Cordova-iOS on Windows.

Closes #971.
Closes #1016.
Closes #1084.