refactor: Define a protocol for scheme handling plugins
We don't rely on the protocol itself being implemented by the plugins
(we continue to check with `-respondsToSelector:`) but this allows us to
avoid `objc_msgSend` and provides a way to document some of this plugin
behaviour that is not otherwise explained.

This should also resolve the unsafe plugin iteration issue that was
mentioned in GH-1272 and GH-1030 by always iterating over an array of
plugin objects that is a copy (due to calling `-allValues`).