fix: CDV.h wasn't bringing in Cordova.h consistently
The fact that there's a circular dependency between the headers isn't
actually a problem in Objective-C because it uses `#import` rather than
`#include` and the compiler ensures each header is only included once.

Somehow that was also causing it to only evaluate the header once and
not differentiate imports where the macro was defined vs not defined.