fix: Avoid showing error page on cancelled navigation
This was previously implemented back in the UIWebView days as CB-13222
in https://github.com/apache/cordova-ios/pull/334 and was then proposed
in https://github.com/apache/cordova-plugin-wkwebview-engine/pull/84 for
the WKWebView plugin, but that didn't get reviewed/merged before the
repo was archived. The fix is still valid, so we'll port it to the
current version of the WKWebView plugin.

Closes GH-959.

Co-Authored-By: Gwyn Judd <gjudd@stayinfront.com>