feat!: Better Catalyst build support (#1313)
* feat!: Support for Catalyst builds

`cordova build ios --device --target=mac ...`

Output is to `build/Debug-maccatalyst`

* fix: Disable iPad-on-Mac builds in favour of Catalyst

Xcode warns that only one can be enabled at a time, so we need to pick
one, and Catalyst seems like the better option for a proper macOS app
experience.

* feat!: Enable compiling for Apple Vision platform

* chore(ci): Add unit tests for run and Catalyst stuff

* fix(build): Don't check for ios-deploy at build time

If neither `--device` nor `--emulator` are specified for the build
command, it will check for a connected device and assume `--device` if
one is found. However, it was also checking for the availability of the
ios-deploy tool which is used to deploy to a connected device.

If we're just building, we don't need to check for a deploy tool. The
run command already has this check to ensure that ios-deploy is
available before actually trying to deploy.

Closes GH-420.
Closes GH-677.