refactor!(SPM): rename Swift package from CordovaLib to Cordova
This ends up being less confusing for everyone because the resulting
framework is named Cordova.framework and the header paths are all
<Cordova/*.h> but currently you need to reference the Swift dependency
as "CordovaLib" rather than "Cordova".

As far as I can tell from a quick search of public Package.swift files
on GitHub, nobody has actually consumed this as a Swift package, so any
breakage should be minimal and we should take the opportunity to do this
right as part of the major version bump.