fix(Swift): Nullability stuff broke Swift compatibility
CDVPlugin's commandDelegate is a weak pointer, which means technically
in Swift it should be an optional type that requires unwrapping. For
some reason, it is not.

If the wrap the CDVPlugin class in the ASSUME_NONNULL macro, Swift
suddenly starts enforcing that it's an optional, and this breaks all
existing Swift plugins.

You aren't allowed to combine `weak` and `nonnull`, and all the
properties in CDVPlugin are weak, so just... don't wrap it in
ASSUME_NONNULL to make life easier for everyone 🙃