refactor(template): Copy CordovaLib into a packages folder
Pointing a relative path to CordovaLib in the node_modules folder means
that the Xcode project is no longer self-contained, which is significant
change in terms of assumptions people can make about the project files
in CI environments (i.e., preparing all platforms once, then copying
just the platform folders to different runners to build).

So we now copy the CordovaLib folder into a platforms/ios/packages
directory and point to it there.

As a side effect, it seemed worth bringing back the `--link` option to
continue pointing to node_modules for easier development.