fix: Prevent recursion in scrollView monkeypatch
If you call scrollView on a UIView that does not have its own scrollView
method, this hijack method will keep recursively calling itself until
you run out of stack space. Try to guard against that by making sure
that it's not getting called on the same instance multiple times.

TODO: Think about thread safety, but in theory because this is
view-related stuff it's only safe to run on the main thread.