fix(cache): Don't load local data from cache
I'm hoping this will resolve the underlying problem from #1451 & #1471
with a smaller scope of change. In this case, we only set the cache
policy to ignore the local cache when we know that it's being loaded
from the filesystem with a custom app scheme.

We can't override the cache policy for the file scheme, and it seems
better to use the existing default cache policy for things being loaded
over http or https schemes.

Co-Authored-By: Brett Blashko <brett.blashko@kanoapps.com>