fix: Add preference for WKWebView crash recovery behaviour (#1582)
* fix: Add preference for crash recovery behaviour

This adds a `CrashRecoveryBehavior` preference with options `"refresh"`
(default - try to refresh the WebView URL) and `"reload"` (load the
starting page again in the WebView).

This doesn't solve the root issue where sometimes refreshing the WebView
still results in a blank screen, but gives people the option to force a
full reload (which seems to be what people are using as a workaround for
the blank screen bug).

Closes GH-1232.
Closes GH-1235.
Closes GH-1355.
Closes GH-1393.
Closes GH-1533.
Closes GH-1552.

* test(webview): API test for crash recovery behaviour